moved functions to Common methods
